+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/calendar/Day.java
@@ -3,6 +3,9 @@ package com.testament.veltahleon.model.entities.calendar;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.testament.veltahleon.model.entities.history.Language;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import org.hibernate.annotations.Cascade;
 
@@ -10,7 +13,8 @@ import java.util.Objects;
 
 @Entity
 @Table(name = "days")
-@Data
+@Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,6 +26,9 @@ public class Day {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @NotNull(message = "Day name cannot be null!")
+    @NotBlank(message = "Day name cannot be blank!")
+    @NotEmpty(message = "Day name cannot be empty!")
     @Column(name = "name")
     private String name;
 
@@ -38,9 +45,20 @@ public class Day {
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinColumn(name = "language_id", referencedColumnName = "id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"}) //to avoid issues with lazy loading related to ManyToOne relationships, as it would require eager loading instead
     private Language language;
 
     @Column(name = "description", columnDefinition = "text")
     private String description;
+
+//    @Override
+//    public String toString() {
+//        return "Day{" +
+//                "id=" + id +
+//                ", name='" + name + '\'' +
+//                ", dayNumber=" + dayNumber +
+//                ", language=" + language +
+//                ", description='" + description + '\'' +
+//                '}';
+//    }
 }