+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/history/Letter.java
@@ -1,29 +1,70 @@
 package com.testament.veltahleon.model.entities.history;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 
 @Entity
 @Table(name = "letters")
-@Data
+@Getter
+@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode
+@ToString
 public class Letter {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @NotNull(message = "Letter name cannot be null!")
+    @NotBlank(message = "Letter name cannot be blank!")
+    @NotEmpty(message = "Letter name cannot be empty!")
     private String name;
 
-    @ManyToOne
-    @JoinColumn(name = "language_id", referencedColumnName = "id")
-    private Language language;
+
+//    @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+//            CascadeType.DETACH,
+//            CascadeType.MERGE,
+//            CascadeType.PERSIST})
+//    @JoinColumn(name = "language_id", referencedColumnName = "id")
+//    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+        CascadeType.DETACH,
+        CascadeType.MERGE,
+        CascadeType.PERSIST})
+    @JoinTable(name = "letters_languages", joinColumns = @JoinColumn(name = "letter_id"),
+        inverseJoinColumns = @JoinColumn(name = "language_id"))
+    private List<Language> languages;
 
     @Column(name = "script_URL")
     private String scriptURL;
+
+//    @Override
+//    public String toString() {
+//        return "Letter{" +
+//                "id=" + id +
+//                ", name='" + name + '\'' +
+//                ", language=" + language +
+//                ", scriptURL='" + scriptURL + '\'' +
+//                '}';
+//    }
+
+    //Convenience Methods
+    public void addLanguage(Language language) {
+        if(languages == null) {
+            languages = new ArrayList<>();
+        }
+        languages.add(language);
+    }
 }