+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/politics/NationLeader.java
@@ -1,12 +1,13 @@
 package com.testament.veltahleon.model.entities.politics;
 
 import com.testament.veltahleon.abstraction.Leader;
+import com.testament.veltahleon.model.entities.calendar.Year;
 import com.testament.veltahleon.model.entities.places.Nation;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.util.Set;
+import java.util.List;
 
 @EqualsAndHashCode(callSuper = true)
 @Entity
@@ -24,6 +25,10 @@ public class NationLeader extends Leader {
     @OneToOne(mappedBy = "nationalLeader")
     private Nation nation;
 
+    @OneToMany
+    @Column(name = "year")
+    private List<Year> yearBeginningAndEnd;
+
     @OneToMany(mappedBy = "suzerain")
-    private Set<Vassal> vassals;
+    private List<Vassal> vassals;
 }