+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/politics/military/Battle.java
@@ -1,6 +1,6 @@
 package com.testament.veltahleon.model.entities.politics.military;
 
-import java.util.Map;
+import java.util.List;
 
 import com.testament.veltahleon.model.entities.calendar.Year;
 import com.testament.veltahleon.model.entities.politics.NationLeader;
@@ -23,31 +23,52 @@ public class Battle {
     @Column(name = "battle_name")
     private String battleName;
 
+//    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+//            CascadeType.DETACH,
+//            CascadeType.MERGE,
+//            CascadeType.PERSIST})
+//    @JoinTable(name = "nation_leaders_battles", joinColumns = @JoinColumn(name = "nation_leader_id"),
+//            inverseJoinColumns = @JoinColumn(name = "battle_id"))
+////    @ElementCollection(fetch = FetchType.LAZY)
+////    @CollectionTable(name = "proponent_opponent_nation_leaders_battles", joinColumns = @JoinColumn(name = "battle_id"))
+////    @MapKeyColumn(name = "proponent_or_opponent_nation_leaders")
+//    @MapKeyColumn(name = "nation_leaders_proponent_and_opponent")
+//    @Column(name = "nation_leaders")
+//    private Map<String, NationLeader> nationalLeaders;
+
     @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
             CascadeType.DETACH,
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinTable(name = "nation_leaders_battles", joinColumns = @JoinColumn(name = "nation_leader_id"),
             inverseJoinColumns = @JoinColumn(name = "battle_id"))
-//    @ElementCollection(fetch = FetchType.LAZY)
-//    @CollectionTable(name = "proponent_opponent_nation_leaders_battles", joinColumns = @JoinColumn(name = "battle_id"))
-//    @MapKeyColumn(name = "proponent_or_opponent_nation_leaders")
-    @MapKeyColumn(name = "nation_leaders_proponent_and_opponent")
-    @Column(name = "nation_leaders")
-    private Map<String, NationLeader> nationalLeaders;
+    private List<NationLeader> nationalLeaders;
 
-    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
-            CascadeType.DETACH,
-            CascadeType.MERGE,
-            CascadeType.PERSIST})
-    @JoinTable(name = "army_leaders_battles", joinColumns = @JoinColumn(name = "army_leader_id"),
-            inverseJoinColumns = @JoinColumn(name = "battle_id"))
-//    @ElementCollection(fetch = FetchType.LAZY)
-//    @CollectionTable(name = "proponent_opponent_military_leaders_battles", joinColumns = @JoinColumn(name = "battle_id"))
-//    @MapKeyColumn(name = "proponent_or_opponent_military_leaders")
-    @MapKeyColumn(name = "army_leaders_proponent_and_opponent")
-    @Column(name = "army_leaders")
-    private Map<String, ArmyLeader> armyLeaders;
+//    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+//            CascadeType.DETACH,
+//            CascadeType.MERGE,
+//            CascadeType.PERSIST})
+//    @JoinTable(name = "army_leaders_battles", joinColumns = @JoinColumn(name = "army_leader_id"),
+//            inverseJoinColumns = @JoinColumn(name = "battle_id"))
+////    @ElementCollection(fetch = FetchType.LAZY)
+////    @CollectionTable(name = "proponent_opponent_military_leaders_battles", joinColumns = @JoinColumn(name = "battle_id"))
+////    @MapKeyColumn(name = "proponent_or_opponent_military_leaders")
+//    @MapKeyColumn(name = "army_leaders_proponent_and_opponent")
+//    @Column(name = "army_leaders")
+//    private Map<String, ArmyLeader> armyLeaders;
+
+//    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+//            CascadeType.DETACH,
+//            CascadeType.MERGE,
+//            CascadeType.PERSIST})
+//    @JoinTable(name = "armies_battles", joinColumns = @JoinColumn(name = "army_id"),
+//            inverseJoinColumns = @JoinColumn(name = "battle_id"))
+////    @ElementCollection(fetch = FetchType.LAZY)
+////    @CollectionTable(name = "proponent_opponent_armies_battles", joinColumns = @JoinColumn(name = "battle_id"))
+////    @MapKeyColumn(name = "proponent_or_opponent_armies")
+//    @MapKeyColumn(name = "armies_proponent_and_opponent")
+//    @Column(name = "armies")
+//    private Map<String, Army> armies;
 
     @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
             CascadeType.DETACH,
@@ -55,12 +76,7 @@ public class Battle {
             CascadeType.PERSIST})
     @JoinTable(name = "armies_battles", joinColumns = @JoinColumn(name = "army_id"),
             inverseJoinColumns = @JoinColumn(name = "battle_id"))
-//    @ElementCollection(fetch = FetchType.LAZY)
-//    @CollectionTable(name = "proponent_opponent_armies_battles", joinColumns = @JoinColumn(name = "battle_id"))
-//    @MapKeyColumn(name = "proponent_or_opponent_armies")
-    @MapKeyColumn(name = "armies_proponent_and_opponent")
-    @Column(name = "armies")
-    private Map<String, Army> armies;
+    private List<Army> armies;
 
     @OneToOne
     @JoinColumn(name = "year_id")