+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/religion/mythology/Pantheon.java
@@ -1,13 +1,12 @@
 package com.testament.veltahleon.model.entities.religion.mythology;
 
-import com.testament.veltahleon.model.entities.religion.mythology.Deity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Set;
+import java.util.List;
 
 @Entity
 @Table(name = "pantheons")
@@ -23,5 +22,5 @@ public class Pantheon {
 
     @OneToMany
     @JoinColumn(name = "pantheon_id")
-    private Set<Deity> deities;
+    private List<Deity> deities;
 }