+++ veltahleon/src/main/java/com/testament/veltahleon/repositories/repo/spring/boot/data/jpa/repository/ifc/history/LetterRepository.java
@@ -1,9 +1,14 @@
 package com.testament.veltahleon.repositories.repo.spring.boot.data.jpa.repository.ifc.history;
 
 import com.testament.veltahleon.model.entities.history.Letter;
-import com.testament.veltahleon.model.entities.history.library.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Collection;
 
-public interface LetterRepository extends JpaRepository<Letter, Long> {}
+@Repository
+public interface LetterRepository extends JpaRepository<Letter, Long> {
+    Letter findByName(String name);
+    //Collection<Letter> findByLanguageName(String languageName);
+    long countByName(String name);
+}