+++ veltahleon/src/main/java/com/testament/veltahleon/services/entities/repo/implementation/history/LanguageServiceImpl.java
@@ -3,9 +3,11 @@ package com.testament.veltahleon.services.entities.repo.implementation.history;
 import com.testament.veltahleon.model.entities.history.Language;
 import com.testament.veltahleon.repositories.repo.spring.boot.data.jpa.repository.ifc.history.LanguageRepository;
 import com.testament.veltahleon.services.entities.repo.ifc.history.LanguageService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
@@ -19,8 +21,13 @@ public class LanguageServiceImpl implements LanguageService {
     private final LanguageRepository languageRepository;
 
     @Override
+    public Collection<Language> getLanguagesWithPagination(int pageNumber, int numberOfRecords) {
+        return languageRepository.findAll(PageRequest.of(pageNumber, numberOfRecords)).toList();
+    }
+
+    @Override
     public Collection<Language> getLanguages() {
-        return null;
+        return languageRepository.findAll();
     }
 
     @Override
@@ -34,16 +41,20 @@ public class LanguageServiceImpl implements LanguageService {
     }
 
     @Override
+    @Transactional
     public Boolean deleteLanguageByID(Long id) {
-        return null;
+        languageRepository.deleteById(id);
+        return Boolean.TRUE;
     }
 
     @Override
+    @Transactional
     public Language saveLanguage(Language language) {
-        return null;
+        return languageRepository.save(language);
     }
 
     @Override
+    @Transactional
     public Language updateLanguage(Language language) {
         return null;
     }